requirements.txt
Flask
Flask-MySQLdb
pyyaml